+++ neuron.py
@@ -1,6 +1,3 @@
-import random
-
-
 class Neuron:
     def __init__(self, value=0):
         self.value = value
